fix chimera_da_push_sized_buf to remove warning

+++ src/internals/da.h
@@ -20,44 +20,44 @@
 
 #define chimera_da_get(da, i) (assert(i < chimera_da_len(da)), da[i])
 
-#define chimera_da_push(da, item)                                              \
-  do {                                                                         \
-    if ((da)->count >= (da)->cap) {                                            \
-      (da)->cap = (da)->cap == 0 ? CHIMERA_INIT_DA_CAP : (da)->cap * 2;        \
-      (da)->items =                                                            \
-          CHIMERA_DA_REALLOC((da)->items, (da)->cap * sizeof(*(da)->items));      \
-      assert((da)->items != NULL && "You broke or what?");                     \
-    }                                                                          \
-                                                                               \
-    (da)->items[(da)->count++] = (item);                                       \
+#define chimera_da_push(da, item)                                             \
+  do {                                                                        \
+    if ((da)->count >= (da)->cap) {                                           \
+      (da)->cap = (da)->cap == 0 ? CHIMERA_INIT_DA_CAP : (da)->cap * 2;       \
+      (da)->items =                                                           \
+          CHIMERA_DA_REALLOC((da)->items, (da)->cap * sizeof(*(da)->items));  \
+      assert((da)->items != NULL && "You broke or what?");                    \
+    }                                                                         \
+                                                                              \
+    (da)->items[(da)->count++] = (item);                                      \
   } while (0)
 
-#define chimera_da_push_sized_buf(da, buf, buf_sz)                             \
-  do {                                                                         \
-    for (int i = 0; i < buf_sz; ++i) {                                         \
-      chimera_da_push(da, buf[i]);                                             \
-    }                                                                          \
+#define chimera_da_push_sized_buf(da, buf, buf_sz)\
+  do {                                            \
+    for (size_t i = 0; i < buf_sz; ++i) {         \
+      chimera_da_push(da, buf[i]);                \
+    }                                             \
   } while (0)
 
-#define chimera_da_push_buf(da, buf)                                           \
+#define chimera_da_push_buf(da, buf)              \
   chimera_da_push_sized_buf(da, buf, strlen(buf))
 
-#define chimera_da_push_mult(da, bufs, bufs_c)                                 \
-  do {                                                                         \
-    if ((da)->count + bufs_c > (da)->cap) {                                    \
-      if ((da)->cap == 0) {                                                    \
-        (da)->cap = CHIMERA_INIT_DA_CAP;                                       \
-      }                                                                        \
-      while ((da)->count + bufs_c > (da)->cap) {                               \
-        (da)->cap *= 2;                                                        \
-      }                                                                        \
-      (da)->items =                                                            \
-          CHIMERA_DA_REALLOC((da)->items, (da)->cap * sizeof(*(da)->items));      \
-      assert((da)->items != NULL);                                             \
-    }                                                                          \
-    memcpy((da)->items + (da)->count, (bufs),                                  \
-           (bufs_c) * sizeof(*(da)->items));                                   \
-    (da)->count += (bufs_c);                                                   \
+#define chimera_da_push_mult(da, bufs, bufs_c)                                \
+  do {                                                                        \
+    if ((da)->count + bufs_c > (da)->cap) {                                   \
+      if ((da)->cap == 0) {                                                   \
+        (da)->cap = CHIMERA_INIT_DA_CAP;                                      \
+      }                                                                       \
+      while ((da)->count + bufs_c > (da)->cap) {                              \
+        (da)->cap *= 2;                                                       \
+      }                                                                       \
+      (da)->items =                                                           \
+          CHIMERA_DA_REALLOC((da)->items, (da)->cap * sizeof(*(da)->items));  \
+      assert((da)->items != NULL);                                            \
+    }                                                                         \
+    memcpy((da)->items + (da)->count, (bufs),                                 \
+           (bufs_c) * sizeof(*(da)->items));                                  \
+    (da)->count += (bufs_c);                                                  \
   } while (0)
 
 typedef struct {
